# blog/openapi.yaml
openapi: 3.0.0
info:
    title: Blog API
    version: v1
    description: Save and load content from the file system.

externalDocs:
    description: Github
    url: https://github.com/abaldeweg/warehouse-server

servers:
    - url: http://localhost:8080/

paths:
    /home:
        get:
            summary: Get a list of all articles
            description: Retrieves a list of all articles with their slug, date and a short summary.
            responses:
                "200":
                    description: Index retrieved successfully.
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                                    properties:
                                        slug:
                                            type: string
                                            description: The slug of the article.
                                        date:
                                            type: string
                                            format: date-time
                                            description: The date of the article.
                                        summary:
                                            type: string
                                            description: A short summary of the article.
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
            security:
                - APIKeyAuth: []

    /home/new/{days}:
        get:
            summary: Get the number of new articles
            description: Retrieves the number of articles newer than the specified number of days.
            parameters:
                - in: path
                  name: days
                  schema:
                      type: integer
                  required: true
                  description: The number of days to look back for new articles.
                  example: 7
            responses:
                "200":
                    description: Number of new articles retrieved successfully.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    new_articles:
                                        type: integer
                                        description: The number of new articles.
                "400":
                    description: Bad Request - Invalid 'days' parameter. Days must be an integer.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
            security:
                - APIKeyAuth: []

    /article/{path}:
        get:
            summary: Get content by path
            description: Retrieves content from the filesystem based on the provided path.
            parameters:
                - in: path
                  name: path
                  schema:
                      type: string
                  required: true
                  description: The path to the content file.
                  example: test/test.md
            responses:
                "200":
                    description: Content retrieved successfully.
                    content:
                        text/plain:
                            schema:
                                type: string
                "500":
                    description: Internal server error.
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    error:
                                        type: string
            security:
                - APIKeyAuth: []

components:
    securitySchemes:
        APIKeyAuth:
            type: apiKey
            in: header
            name: X-API-Key